+++ app/src/main/java/com/github/libretube/ui/preferences/AdvancedSettings.kt
@@ -1,48 +1,17 @@
 package com.github.libretube.ui.preferences
 
-import android.net.Uri
 import android.os.Bundle
-import androidx.activity.result.ActivityResultLauncher
-import androidx.activity.result.contract.ActivityResultContracts
-import androidx.activity.result.contract.ActivityResultContracts.CreateDocument
 import androidx.preference.ListPreference
 import androidx.preference.Preference
 import com.github.libretube.R
 import com.github.libretube.constants.PreferenceKeys
-import com.github.libretube.obj.BackupFile
 import com.github.libretube.ui.activities.SettingsActivity
 import com.github.libretube.ui.base.BasePreferenceFragment
-import com.github.libretube.ui.dialogs.BackupDialog
-import com.github.libretube.util.BackupHelper
 import com.github.libretube.util.ImageHelper
 import com.github.libretube.util.PreferenceHelper
 import com.google.android.material.dialog.MaterialAlertDialogBuilder
-import java.time.LocalDate
-import java.time.LocalTime
 
 class AdvancedSettings : BasePreferenceFragment() {
-
-    // backup and restore database
-    private lateinit var getBackupFile: ActivityResultLauncher<String>
-    private lateinit var createBackupFile: ActivityResultLauncher<String>
-    private var backupFile = BackupFile()
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        getBackupFile =
-            registerForActivityResult(
-                ActivityResultContracts.GetContent()
-            ) { uri: Uri? ->
-                BackupHelper(requireContext()).restoreAdvancedBackup(uri)
-            }
-
-        createBackupFile = registerForActivityResult(
-            CreateDocument("application/json")
-        ) { uri: Uri? ->
-            BackupHelper(requireContext()).advancedBackup(uri, backupFile)
-        }
-
-        super.onCreate(savedInstanceState)
-    }
 
     override fun onCreatePreferences(savedInstanceState: Bundle?, rootKey: String?) {
         setPreferencesFromResource(R.xml.advanced_settings, rootKey)
@@ -59,22 +28,6 @@ class AdvancedSettings : BasePreferenceFragment() {
         val resetSettings = findPreference<Preference>(PreferenceKeys.RESET_SETTINGS)
         resetSettings?.setOnPreferenceClickListener {
             showResetDialog()
-            true
-        }
-
-        val advancesBackup = findPreference<Preference>("backup")
-        advancesBackup?.setOnPreferenceClickListener {
-            BackupDialog {
-                backupFile = it
-                createBackupFile.launch(getBackupFileName())
-            }
-                .show(childFragmentManager, null)
-            true
-        }
-
-        val restoreAdvancedBackup = findPreference<Preference>("restore")
-        restoreAdvancedBackup?.setOnPreferenceClickListener {
-            getBackupFile.launch("application/json")
             true
         }
     }
@@ -94,10 +47,5 @@ class AdvancedSettings : BasePreferenceFragment() {
                 activity?.recreate()
             }
             .show()
-    }
-
-    private fun getBackupFileName(): String {
-        val time = LocalTime.now().toString().split(".").firstOrNull()
-        return "libretube-backup-${LocalDate.now()}-$time.json"
     }
 }
