+++ app/src/main/res/values/strings.xml
@@ -403,6 +403,10 @@
     <string name="double_tap_seek_summary">Tap twice at the left or right to rewind or forward the player position.</string>
     <string name="all_caught_up">You\'re all caught up</string>
     <string name="all_caught_up_summary">You\'ve seen all new videos</string>
+    <string name="import_playlists">Import playlists</string>
+    <string name="export_playlists">Export playlists</string>
+    <string name="app_backup">App Backup</string>
+    <string name="backup_restore_summary">Import &amp; export subscriptions, playlists, ...</string>
 
     <!-- Notification channel strings -->
     <string name="download_channel_name">Download Service</string>
