+++ app/src/main/res/xml/import_export_settings.xml
@@ -1,0 +1,49 @@
+<?xml version="1.0" encoding="utf-8"?>
+<PreferenceScreen xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto">
+
+    <PreferenceCategory app:title="@string/subscriptions">
+
+        <Preference
+            android:icon="@drawable/ic_download_filled"
+            android:summary="@string/import_from_yt_summary"
+            app:key="import_from_yt"
+            app:title="@string/import_from_yt" />
+
+        <Preference
+            android:icon="@drawable/ic_upload"
+            app:key="export_subs"
+            app:title="@string/export_subscriptions" />
+
+    </PreferenceCategory>
+
+    <PreferenceCategory app:title="@string/playlists">
+
+        <Preference
+            android:icon="@drawable/ic_download_filled"
+            android:summary="@string/import_playlists"
+            app:key="import_from_yt"
+            app:title="@string/import_from_yt" />
+
+        <Preference
+            android:icon="@drawable/ic_upload"
+            app:key="export_subs"
+            app:title="@string/export_playlists" />
+
+    </PreferenceCategory>
+
+    <PreferenceCategory app:title="@string/app_backup">
+
+        <Preference
+            android:icon="@drawable/ic_backup"
+            app:key="backup"
+            app:title="@string/backup" />
+
+        <Preference
+            android:icon="@drawable/ic_restore"
+            app:key="restore"
+            app:title="@string/restore" />
+
+    </PreferenceCategory>
+
+</PreferenceScreen>
