+++ app/src/main/res/xml/settings.xml
@@ -51,6 +51,12 @@
         app:title="@string/notifications" />
 
     <Preference
+        android:icon="@drawable/ic_backup"
+        app:key="backup_restore"
+        app:summary="@string/backup_restore_summary"
+        app:title="@string/backup_restore" />
+
+    <Preference
         android:icon="@drawable/ic_list"
         app:key="advanced"
         app:summary="@string/advanced_summary"
